enable cors for origin and refactor
